requirements.txt
pymongo==3.11.1
python-dotenv==0.15.0
beautifulsoup4==4.9.3
numpy==1.15.4
pandas==1.1.4